Add files via upload
I did a simple run using the original jupyter notebook and got a score of 7268948 with computed reaction energy: -7.18 eV, not sure what was the runtime but probably around 30min. Since the energy seems too far I decided to check if I could improve the energy by a subsequent run but starting from the best parameters of the previous run. So, by defining new initial_point_list to be the best result of the previous run and running for about 37 minutes I obtained total score of 7264727 (decrease by 4221) - not sure why it changed since the circuit was still the same. Now the computed reaction energy was -6.34 eV (an increase by 0.84 eV).